Move the tag related functinality into a single module.
