+++ .eslintrc
@@ -18,7 +18,8 @@
         ]
     },
     "env": {
-        "node": true
+        "node": true,
+        "es6": true
     },
     "extends": "eslint:recommended"
 }
