+++ lib/specifications/files/feature-files/tags.js
@@ -3,6 +3,11 @@
  */
 'use strict';
 
+module.exports = {
+  count: count,
+  filter: filter
+};
+
 /**
  * Add tags to the tag counting object.
  * @param Object tagsCount            The tag counting object.
@@ -95,6 +100,81 @@ function count(featureData, tagCountObject, requestedTag) {
   return tagCountObject;
 }
 
-module.exports = {
-  count: count
-};
+/*
+  Applying filtering based on feature and scenario tags.
+ */
+function filter(projectData, projectTags, currentTags) {
+  var tagNames = [];
+
+  // Count the tags in the project.
+  projectData.files.forEach(function(file) {
+    if (!file.isFeatureFile || file.error) {
+      return;
+    }
+
+    // This counts tags and marks when an
+    // object contains the requested tag.
+    projectTags = count(file.data, projectTags, currentTags);
+  });
+  tagNames = Object.keys(projectTags);
+  projectData.hasTags = !!tagNames.length;
+  // Mark the currently requested tag if any,
+  // this is used to set the selected option
+  // in the tag select box.
+  tagNames.forEach(function(name) {
+    // Currently on one tag is passed in the query parameter.
+    if (name === currentTags) {
+      projectTags[name].isCurrent = true;
+    }
+  });
+  projectData.tags = projectTags;
+
+  // Filter the features and scenarios based on
+  // whether they contain the requested tag.
+  if (currentTags) {
+    projectData.files = projectData.files.filter(function(file) {
+      var feature;
+      var featureScenarioContainsTag = false;
+
+      // Filter out non-feature or erroring files.
+      if (!file.isFeatureFile || file.error) {
+        return false;
+      }
+
+      // If the FEATURE contains the tag keep it and take no further action.
+      feature = file.data;
+      if (feature.containsRequestedTag) {
+        return true;
+      }
+
+      // Decide whether to filter a scenario based on whether it or a child
+      // object has the requested tag.
+      feature.scenarioDefinitions.forEach(function (scenario, index, defs) {
+
+        // if any example contains the tag keep all examples.
+        if (scenario.type === 'ScenarioOutline') {
+          scenario.examples.forEach(function(example) {
+            if (example.containsRequestedTag) {
+              scenario.containsRequestedTag = true;
+            }
+          });
+        }
+        if (scenario.containsRequestedTag) {
+          featureScenarioContainsTag = true;
+        } else {
+          // Set the scenario to undefined so it won't be rendered.
+          defs[index] = undefined;
+        }
+      });
+      // Remove undefined scenarios because handlbars' `each`
+      // helper doen't ignore undefined array elements.
+      feature.scenarioDefinitions = feature.scenarioDefinitions.filter(function(scenario) { return scenario !== undefined; });
+
+      // Retain or lose the feature depending on whether a scenario
+      // contained the requested tag.
+      return featureScenarioContainsTag;
+    });
+  }
+
+  return [projectData, projectTags];
+}
