+++ routes/project.js
@@ -11,7 +11,7 @@ var arrrayToTree = require('file-tree');
 var TreeModel = require('tree-model');
 
 var processFiles = require('../lib/specifications/files/process-files');
-var countTags = require('../lib/specifications/files/feature-files/tags').count;
+var filterByTag = require('../lib/specifications/files/feature-files/tags').filter;
 
 var getProject = require('../lib/specifications/projects/project').get;
 var getProjectData = require('../lib/specifications/projects/project').getData;
@@ -25,6 +25,8 @@ var appConfig = require('../lib/configuration/app-config').get();
 function getRender(res, appConfig, renderOptions) {
   return function render(projectData) {
     var renderingData = {};
+
+    // The tags object associated with this project.
     var projectTags = {};
 
     renderingData.openBurgerMenu = renderOptions.openBurgerMenu;
@@ -49,7 +51,7 @@ function getRender(res, appConfig, renderOptions) {
     // Applying views from configuration.
     // Chrome hasn't turned destructuring assignment on yet,
     // so I'm cheating
-    var ret = applyView(projectData, renderingData);
+    let ret = applyView(projectData, renderingData);
     projectData = ret[0];
     renderingData = ret[1];
 
@@ -71,11 +73,9 @@ function getRender(res, appConfig, renderOptions) {
     var promisesForFileContent = projectData.files.map(function(f) {return f.contentPromise;});
     return Promise.all(promisesForFileContent)
       .then(function() {
-        var tagNames = [];
 
         // Mix in the file content.
         projectData.files = projectData.files.map(processFiles.processFileContent);
-
 
         // If the project config contains a URL format
         // for creating links to edit files then grab it.
@@ -93,81 +93,9 @@ function getRender(res, appConfig, renderOptions) {
           });
         }
 
-
-        /*
-          Applying filtering based on feature and scenario tags.
-         */
-
-        var currentTags = renderOptions.currentTags;
-        // Count the tags in the project.
-        projectData.files.forEach(function(file) {
-          if (!file.isFeatureFile || file.error) {
-            return;
-          }
-
-          // This counts tags and marks when an
-          // object contains the requested tag.
-          projectTags = countTags(file.data, projectTags, currentTags);
-        });
-        tagNames = Object.keys(projectTags);
-        projectData.hasTags = !!tagNames.length;
-        // Mark the currently requested tag if any,
-        // this is used to set the selected option
-        // in the tag select box.
-        tagNames.forEach(function(name) {
-          // Currently on one tag is passed in the query parameter.
-          if (name === currentTags) {
-            projectTags[name].isCurrent = true;
-          }
-        });
-        projectData.tags = projectTags;
-
-        // Filter the features and scenarios based on
-        // whether they contain the requested tag.
-        if (currentTags) {
-          projectData.files = projectData.files.filter(function(file) {
-            var feature;
-            var featureScenarioContainsTag = false;
-
-            // Filter out non-feature or erroring files.
-            if (!file.isFeatureFile || file.error) {
-              return false;
-            }
-
-            // If the feature contains the tag keep it and take no
-            // further action.
-            feature = file.data;
-            if (feature.containsRequestedTag) {
-              return true;
-            }
-
-            feature.scenarioDefinitions.forEach(function (scenario, index, defs) {
-
-              // if any example contains the tag keep all examples.
-              if (scenario.type === 'ScenarioOutline') {
-                scenario.examples.forEach(function(example) {
-                  if (example.containsRequestedTag) {
-                    scenario.containsRequestedTag = true;
-                  }
-                });
-              }
-              if (scenario.containsRequestedTag) {
-                featureScenarioContainsTag = true;
-              } else {
-                // Set the scenario to undefined so it won't be rendered.
-                defs[index] = undefined;
-              }
-            });
-            // Remove undefined scenarios because handlbars' `each`
-            // helper doen't ignore undefined array elements.
-            feature.scenarioDefinitions = feature.scenarioDefinitions.filter(function(scenario) { return scenario !== undefined; });
-
-            // Retain or lose the feature depending on whether a scenario
-            // contained the requested tag.
-            return featureScenarioContainsTag;
-          });
-        }
-
+        let ret = filterByTag(projectData, projectTags, renderOptions.currentTags);
+        projectData = ret[0];
+        projectTags = ret[1];
 
         /*
           Generate a tree data structure from the flat file list.
